+++ source/renderer/app/themes/daedalus/incentivized-testnet.js
@@ -120,16 +120,14 @@ export const INCENTIVIZED_TESTNET_THEME_OUTPUT = {
   },
   delegationSetupWizard: {
     '--theme-delegation-steps-activation-steps-indicator-color': '#ffffff',
-    '--theme-delegation-steps-activation-description-color': '#ffffffcc',
-    '--theme-delegation-steps-activation-fees-label-color': '#ffffff',
-    '--theme-delegation-steps-activation-fees-amount-color': '#ea4c5b',
-    '--theme-delegation-steps-activation-address-value-color': '#ffffff',
+    '--theme-delegation-steps-success-description-color': '#ffffff',
+    '--theme-delegation-steps-success-tada-color': 'rgba(255, 255, 255, 1)',
     '--theme-delegation-steps-choose-stake-pool-checkmark-icon-color':
       'rgba(255, 255, 255, 0.2)',
     '--theme-delegation-steps-choose-stake-pool-delegated-pools-label-color':
       '#ffffff',
     '--theme-delegation-steps-choose-stake-pool-ticker-color':
-      'rgba(255, 255, 255, 0.2)',
+      'rgba(255, 255, 255, 1)',
     '--theme-delegation-steps-choose-stake-pool-select-box-placeholder-color':
       'rgba(255, 255, 255, 0.2)',
     '--theme-delegation-steps-choose-stake-pool-selected-checkmark-icon-color':
@@ -140,7 +138,7 @@ export const INCENTIVIZED_TESTNET_THEME_OUTPUT = {
       'rgba(42, 43, 60, 1)',
     '--theme-delegation-steps-choose-stake-pool-thumb-border-color':
       'rgba(255, 255, 255, 0.2)',
-    '--theme-delegation-steps-choose-stake-pool-title-color': '#ffffffcc',
+    '--theme-delegation-steps-choose-stake-pool-title-color': '#ffffff',
     '--theme-delegation-steps-choose-stake-pool-tooltip-arrow-color':
       '#ffffffe6',
     '--theme-delegation-steps-choose-stake-pool-tooltip-background-color':
@@ -705,6 +703,9 @@ export const INCENTIVIZED_TESTNET_THEME_OUTPUT = {
     '--theme-staking-delegation-center-gear-icon-fill-color-active': '#ffffff',
     '--theme-staking-delegation-center-no-wallets-instructions-color':
       '#ffffff',
+    '--theme-staking-delegation-center-divider-border-color':
+      'rgba(255, 255, 255, 0.2)',
+    '--theme-staking-delegation-center-fees-amount-color': '#eb4a22',
     '--theme-staking-countdown-widget-background-color':
       'rgba(255, 255, 255, 0.07)',
     '--theme-staking-countdown-widget-delimeter-background-color': '#ffffff',