+++ source/renderer/app/themes/daedalus/yellow.js
@@ -125,10 +125,8 @@ export const YELLOW_THEME_OUTPUT = {
   },
   delegationSetupWizard: {
     '--theme-delegation-steps-activation-steps-indicator-color': '#2d2d2d',
-    '--theme-delegation-steps-activation-description-color': '#2d2d2dcc',
-    '--theme-delegation-steps-activation-fees-label-color': '#2d2d2d',
-    '--theme-delegation-steps-activation-fees-amount-color': '#ea4c5b',
-    '--theme-delegation-steps-activation-address-value-color': '#2d2d2d',
+    '--theme-delegation-steps-success-description-color': '#2d2d2d',
+    '--theme-delegation-steps-success-tada-color': 'rgba(45, 45, 45, 1)',
     '--theme-delegation-steps-choose-stake-pool-checkmark-icon-color':
       '#e1dac6',
     '--theme-delegation-steps-choose-stake-pool-delegated-pools-label-color':
@@ -142,7 +140,7 @@ export const YELLOW_THEME_OUTPUT = {
       '#fafbfc',
     '--theme-delegation-steps-choose-stake-pool-thumb-background-color': '#fff',
     '--theme-delegation-steps-choose-stake-pool-thumb-border-color': '#e1dac6',
-    '--theme-delegation-steps-choose-stake-pool-title-color': '#2d2d2dcc',
+    '--theme-delegation-steps-choose-stake-pool-title-color': '#2d2d2d',
     '--theme-delegation-steps-choose-stake-pool-tooltip-arrow-color':
       '#2d2d2de6',
     '--theme-delegation-steps-choose-stake-pool-tooltip-background-color':
@@ -709,6 +707,8 @@ export const YELLOW_THEME_OUTPUT = {
     '--theme-staking-delegation-center-gear-icon-fill-color-active': '#2d2d2d',
     '--theme-staking-delegation-center-no-wallets-instructions-color':
       '#2d2d2d',
+    '--theme-staking-delegation-center-divider-border-color': '#e1dac6',
+    '--theme-staking-delegation-center-fees-amount-color': '#ea4c5b',
     '--theme-staking-countdown-widget-background-color':
       'rgba(45, 45, 45, 0.07)',
     '--theme-staking-countdown-widget-delimeter-background-color': '#2d2d2d',